+++ MHD/C_src/GTK/tree_view_viewer_colormap.c
@@ -55,7 +55,7 @@ void init_colormap_params_4_viewer(struct kemoviewer_type *kemo_sgl,
 
 void load_color_opacity_map_from_list(struct psf_menu_val *psf_current_menu,
 			struct colormap_view *color_vws){
-	int icomp = psf_current_menu->icomp_draw_psf;
+	long icomp = psf_current_menu->icomp_draw_psf;
 	dup_real2_clist(color_vws->cmap_vws->r2_clist_gtk,
 					psf_current_menu->cmap_psf_comp[icomp]->colormap);
 	dup_real2_clist(color_vws->opacity_vws->r2_clist_gtk,
