+++ MHD/C_src/KEMO_GL/draw_map_4_PSF.h
@@ -27,7 +27,7 @@ int check_draw_map(struct kemo_array_control *psf_a);
 void set_map_PSF_isolines_buffer(struct psf_data **psf_s, struct psf_menu_val **psf_m,
                                  struct kemo_array_control *psf_a, struct view_element *view_s,
                                  struct gl_strided_buffer *mline_buf);
-void set_map_patch_buffer(int ist_psf, int ied_psf, struct psf_data **psf_s,
+void set_map_patch_buffer(long ist_psf, long ied_psf, struct psf_data **psf_s,
                           struct psf_menu_val **psf_m,
                           struct kemo_array_control *psf_a,
                           struct gl_strided_buffer *map_buf);
