+++ MHD/C_src/KEMO_GL/m_kemoview_psf.c
@@ -171,7 +171,7 @@ int get_each_PSF_field_param(int selected, struct kemoview_psf *kemo_psf){
 	} else if(selected == COMPONENT_SEL_FLAG){
 		output =  send_draw_comp_id_psf(kemo_psf->psf_m[i_current]);
 	} else if(selected == DRAW_ADDRESS_FLAG){
-		output =  send_draw_component_psf(kemo_psf->psf_m[i_current]);
+		output =  (int) send_draw_component_psf(kemo_psf->psf_m[i_current]);
 	} else if(selected == COORDINATE_FLAG){
 		output =  send_coordinate_id_psf(kemo_psf->psf_d[i_current], 
                                          kemo_psf->psf_m[i_current]);
