+++ MHD/C_src/KEMO_GL/set_each_isoline_to_buf.h
@@ -17,13 +17,13 @@
 /* prototypes */
 
 long add_each_isoline_npatch(const long ist_patch, double v_line,
-                             int icomp, struct psf_data *psf_s);
+                             long icomp, struct psf_data *psf_s);
 long set_each_isoline_to_buf(const long ist_patch, double width,
-                             double v_line, int icomp, double *f_color,
+                             double v_line, long icomp, double *f_color,
                              struct psf_data *psf_s,
                              struct gl_strided_buffer *strided_buf);
 long set_each_map_isoline_to_buf(const long ist_patch, double width, 
-                                 double v_line, int icomp, double *f_color, 
+                                 double v_line, long icomp, double *f_color, 
                                  struct psf_data *psf_s, 
                                  struct gl_strided_buffer *strided_buf);
