+++ MHD/C_src/KEMO_GL/set_each_psf_parameters.h
@@ -37,7 +37,7 @@ void send_each_psf_data_name(struct psf_data *psf_d, struct kv_string *colorname
 
 int send_field_draw_each_psf(struct psf_menu_val *psf_menu);
 int send_draw_comp_id_psf(struct psf_menu_val *psf_menu);
-int send_draw_component_psf(struct psf_menu_val *psf_menu);
+long send_draw_component_psf(struct psf_menu_val *psf_menu);
 int send_coordinate_id_psf(struct psf_data *psf_d, struct psf_menu_val *psf_menu);
 
 void set_texture_psf_from_bgra(struct kemo_array_control *psf_a,
