+++ MHD/C_src/Metal/KemoViewerMetalView.m
@@ -178,7 +178,7 @@
 {
     kemoview_set_view_integer(ISET_ROTATE_AXIS, (int) rotationaxis, kemo_sgl);
     kemoview_set_view_integer(ISET_ROTATE_INCREMENT, (int) int_degree, kemo_sgl);
-    kemoview_step_viewmatrix(istep_quilt, kemo_sgl);
+    kemoview_step_viewmatrix((int) istep_quilt, kemo_sgl);
 
     [self setNeedsDisplay: YES];
     return self;
